# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified Messaging Gateway",
    "description": "Single point of contact for all messaging integrations. Abstracts WhatsApp, Telegram, Email, and SMS providers behind a unified API.",
    "version": "0.4.0"
  },
...
  "paths": {
...
    "/api/v1/accounts/{id}/connect": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "Start the connection process for a self-auth provider (e.g. Baileys). After calling this, poll GET /api/v1/accounts/:id to retrieve the QR code.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "alias": {
                      "type": "string"
                    },
                    "channel": {
                      "type": "string",
                      "enum": [
                        "whatsapp",
                        "telegram",
                        "email",
                        "sms"
                      ]
                    },
                    "provider": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "suspended",
                        "auth_expired",
                        "error",
                        "unchecked"
                      ]
                    },
                    "identity": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "connection": {
                      "type": "object",
                      "description": "Live connection info for providers that manage their own connection (e.g. Baileys). Absent for API-key providers.",
                      "properties": {
                        "managed": {
                          "type": "boolean",
                          "description": "Whether this provider uses managed connections"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "disconnected",
                            "connecting",
                            "connected"
                          ]
                        },
                        "qr": {
                          "type": "string",
                          "description": "QR code data string (render with a QR library to scan)"
                        }
                      },
                      "required": [
                        "managed",
                        "status"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "owner": {
                          "type": "string"
                        },
                        "environment": {
                          "type": "string"
                        },
                        "webhookPath": {
                          "type": "string"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "id",
                    "alias",
                    "channel",
                    "provider",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{id}/pair": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "Request a pairing code for authentication. Enter this code in WhatsApp > Linked Devices > Link with phone number.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number with country code (e.g. \"+34600000001\"). If omitted, uses the number from account identity."
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string"
                    },
                    "pairingCode": {
                      "type": "string",
                      "description": "Enter this code in the messaging app to link"
                    }
                  },
                  "required": [
                    "accountId",
                    "pairingCode"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{id}/disconnect": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "description": "Disconnect a self-auth provider account and clear its session",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "accountId",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}